proof of concept
Cooked this up while watching Sea Eagles v Roosters in the Rugby League preliminary finals with a BBQ chicken pizza from Crust.

This program listens for MAC addresses on a network interface and sends a list (JSON) of unique IP and hostnames to a TCP socket for a Node-Red powered HTML5 dashboard.